5.3更新内容
【功能新增】
1.新增业务应用 文档管理系统
文件夹管理：创建、更名、删除、复制、移动、授权
文档管理：上传、下载、更名、复制、移动 、预览、分享
权限控制：按组织机构和按用户组两种授权模式
在线预览：无需下载，上百种格式文件在线预览
收藏夹：支持将文件夹和文档加入收藏、查看和移除
全文搜索：对文本类、office文档和pdf文档等进行全文搜索
【优化调整】
1.文件上传组件增加失败异常友好信息展示
【bug修复】
1.实体配置模块控制器类模板遗留使用模型属性编码（code）引入服务类的问题，修正为使用模型属性对应的实体编码（entityCode）。

+++ platform-web/src/modules/support/view/attachment/attachmentUploader.vue
@@ -10,6 +10,7 @@
     @file-complete="fileComplete"
     @file-success="fileSuccess"
     @file-added="fileAdded"
+    @file-error="fileError"
     class="w-full"
   >
     <uploader-drop class="uploader-dragover">
@@ -79,6 +80,18 @@ export default {
       type: String,
       default: '/support/attachment/uploadChunk',
       required: false
+    },
+    mergeChunkApi: {
+      type: String,
+      default: 'this.$api.support.attachment',
+      required: false
+    },
+    appendData: {
+      type: Object,
+      required: false,
+      default() {
+        return {}
+      }
     }
   },
   data() {
@@ -139,7 +152,7 @@ export default {
     },
     fileSuccess(rootFile, file) {
       if (file.chunks.length > 1) {
-        //分块上传
+        //构造文件信息
         const param = {
           identifier: file.uniqueIdentifier,
           filename: file.name,
@@ -147,16 +160,31 @@ export default {
           entityType: this.entityType,
           entityId: this.entityId,
           type: file.fileType,
-          totalSize: file.size
+          totalSize: file.size,
+          ...this.appendData
         }
+
         // 合并文件块
-        this.$api.support.attachment.mergeChunks(param).then(() => {
-          // 移除已上传成功的文件
-          this.$refs.uploader.uploader.removeFile(file)
-        })
+        // eslint-disable-next-line no-eval
+        eval(this.mergeChunkApi)
+          .mergeChunks(param)
+          .then(() => {
+            if (!this.showSuccessFiles) {
+              // 移除已上传成功的文件
+              this.$refs.uploader.uploader.removeFile(file)
+            }
+          })
       } else {
-        // 不分块，移除已上传成功的文件
-        this.$refs.uploader.uploader.removeFile(file)
+        if (!this.showSuccessFiles) {
+          // 不分块，移除已上传成功的文件
+          this.$refs.uploader.uploader.removeFile(file)
+        }
+      }
+    },
+    fileError(rootFile, file, responseData, chunk) {
+      const res = JSON.parse(responseData)
+      if (res && res.message) {
+        this.$message.error(res.message)
       }
     }
   }
